allow for signal handlers that schedule or awaken greenlets
if a poll call is interrupted by a signal, before it resumes with the modified
timeout it should check again that greenlets weren't made runnable by the
signal handler itself. otherwise the newly runnable greenlets will still have
to wait until a network event or the poll's timeout has expired.