fixed issue #4
now the return value of poller.register() has to be passed as another argument
to poller.unregister() so it knows which one is being referred to.

updated tests and docs accordingly.

finally, also took this refactoring opportunity to make the kqueue poller only
apply the eventmask *diff*, rather than unregistering everything and
reregistering what we now need (as the poll and epoll pollers have to).