emulation.patched_context and associated plumbing
patched_context is a context manager function, so it can be called in a
with-statement. its implementation causes emulation to be done on entry
into the context (in __enter__), then removed afterward in __exit__.

with the option local=True, for the duration of the context it will also
unpatch on every switch out of the coro, and re-patch on every entry
back in.

for the implementation of that last part a function was needed that
removes a local trace hook, incoming and outgoing, so I added it. since
they are so similar, I also added removal functions for global trace
hooks and global and local exception handlers.