create emulated Threads on demand
when current_thread() is called on a patched threading module, but the
current greenlet was created through direct greenhouse APIs, we have to
create them on the fly in order to give unique and relevant thread ids.
since the greenlet is clearly running at this time, we go ahead and put
the thread in the 'running' state.